qcloud/scripts/requirements.txt
Jinja2==2.10
qcloudapi-sdk-python==2.0.12
pyyaml
